+++ types/rox-browser/v4/index.d.ts
@@ -1,7 +1,6 @@
-// Type definitions for rox-browser 5.0
+// Type definitions for rox-browser 4.8
 // Project: https://rollout.io
 // Definitions by: g-guirado <https://github.com/g-guirado>
-//                 rollout: <https://github.com/rollout>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 // TypeScript Version: 3.0
 
@@ -13,7 +12,7 @@
  */
 
 export interface RoxContainer {
-  [key: string]: Flag | RoxNumber | RoxString;
+  [key: string]: Flag | Configuration<any> | Variant;
 }
 
 /**
@@ -22,8 +21,6 @@ export interface RoxContainer {
  * https://support.rollout.io/docs/javascript-browser-api#section-register
  */
 export function register(namespace: string, roxContainer: RoxContainer): void;
-
-export function register(roxContainer: RoxContainer): void;
 
 /**
  * Set Global Context.
@@ -47,13 +44,11 @@ export interface RoxSetupOptions {
   configurationFetchedHandler?(fetcherResult: RoxFetcherResult): void;
   debugLevel?: 'verbose';
   // https://support.rollout.io/docs/javascript-browser-api#section-using-the-impressionhandler-option
-  impressionHandler?(reporting: RoxReporting, context: unknown): void;
+  impressionHandler?(reporting: RoxReporting, experiment: RoxExperiment, context: unknown): void;
   platform?: string;
   freeze?: RoxFlagFreezeLevel;
   disableNetworkFetch?: boolean;
   devModeSecret?: string;
-  // https://support.rollout.io/docs/javascript-browser-api#section-dynamicPropertyRuleHandler
-  dynamicPropertyRuleHandler?(propName: string, context: any): any;
 }
 
 export enum RoxFetcherStatus {
@@ -61,13 +56,6 @@ export enum RoxFetcherStatus {
   AppliedFromCache = 'APPLIED_FROM_CACHE',
   AppliedFromNetwork = 'APPLIED_FROM_NETWORK',
   ErrorFetchFailed = 'ERROR_FETCH_FAILED'
-}
-
-export enum RoxErrorTrigger {
-  DYNAMIC_PROPERTIES_RULE = 'DYNAMIC_PROPERTIES_RULE',
-  CONFIGURATION_FETCHED_HANDLER = 'CONFIGURATION_FETCHED_HANDLER',
-  IMPRESSION_HANDLER = 'IMPRESSION_HANDLER',
-  CUSTOM_PROPERTY_GENERATOR = 'CUSTOM_PROPERTY_GENERATOR'
 }
 
 export interface RoxFetcherResult {
@@ -80,7 +68,13 @@ export interface RoxFetcherResult {
 export interface RoxReporting {
   name: string;
   value: string;
-  targeting: boolean;
+}
+
+export interface RoxExperiment {
+  identifier: string; //  experiment id
+  name: string;
+  isArchived: boolean;
+  labels: string[]; // experiment's labels. assigned from dashboard
 }
 
 /**
@@ -90,10 +84,8 @@ export interface RoxReporting {
 export function setCustomNumberProperty(name: string, value: number | ((context?: unknown) => number)): void;
 export function setCustomStringProperty(name: string, value: string | ((context?: unknown) => string)): void;
 export function setCustomBooleanProperty(name: string, value: boolean | ((context?: unknown) => boolean)): void;
-
-// https://support.rollout.io/docs/user-space-error-handler
-export function setUserspaceUnhandledErrorHandler(
-  handler: (errorTrigger: RoxErrorTrigger, error: Error) => void
+export function setDynamicCustomPropertyRule(
+  handler: (propName: string, context: unknown) => number | string | boolean
 ): void;
 
 /**
@@ -126,7 +118,7 @@ export enum RoxOverridesPosition {
 }
 
 /**
- * Default is none
+ * Default is untilForeground
  *
  * https://support.rollout.io/docs/flags-update-flow#section-flag-freeze-level
  */
@@ -161,46 +153,53 @@ export class Flag {
 }
 
 /**
- * Used to create and manage Rollout feature flags that determine different predefined string values
+ * Used to create and manage Rollout feature flags that determine different predefined values
  *
- * https://support.rollout.io/docs/javascript-browser-api#section--rox-string-
+ * https://support.rollout.io/docs/javascript-browser-api#section--rox-variant-
  */
-export class RoxString {
-  constructor(defaultValue: string, options?: ReadonlyArray<string>);
+export class Variant<T extends string = string> {
+  constructor(defaultValue: T, options: ReadonlyArray<T>, name?: string);
 
-  // The name of the RoxString
+  // The name of the Variant
   readonly name: string;
 
-  // Default value of the RoxString
-  readonly defaultValue: string;
+  // Default value of the Variant
+  readonly defaultValue: BasicType<T>;
 
-  // Returns the current value of the RoxString, accounting for value overrides
-  getValue(context?: unknown): string;
+  // Returns the current value of the Variant, accounting for value overrides
+  getValue(context?: unknown): BasicType<T>;
 
-  // Unlock the RoxString value from changes from the last time it was freezed
+  // Unlock the Variant value from changes from the last time it was freezed
   unfreeze(): void;
 }
 
 /**
- * Used to create and manage Rollout feature flags that determine different predefined number values
+ * manages a remote configuration setting with a value of type string, boolean, or number.
+ * The constructor sets the default value for the remote configuration setting
  *
- * https://support.rollout.io/docs/javascript-browser-api#section--rox-number-
+ * https://support.rollout.io/docs/javascript-browser-api#section--rox-configuration-
  */
-export class RoxNumber {
-  constructor(defaultValue: number, options?: ReadonlyArray<number>);
+export class Configuration<T extends number | boolean | string> {
+  constructor(defaultValue: T);
 
-  // The name of the RoxNumber
+  // The name of the Configuration
   readonly name: string;
 
-  // Default value of the RoxNumber
-  readonly defaultValue: number;
+  // Default value of the Configuration
+  readonly defaultValue: BasicType<T>;
 
-  // Returns the current value of the RoxNumber, accounting for value overrides
-  getValue(context?: unknown): number;
+  // Returns the current value of the Configuration, accounting for value overrides
+  getValue(context?: unknown): BasicType<T>;
 
-  // Unlock the RoxNumber value from changes from the last time it was freezed
+  // Unlock the Configuration value from changes from the last time it was freezed
   unfreeze(): void;
 }
+
+/**
+ * Ensure that TypeScript properly types things with a basic type.
+ * For example, if T is true, returnedtype shall be boolean, not true
+ */
+export type BasicType<T> = T extends boolean ? boolean : T extends number ? number : T extends string ? string : never;
 
 /**
  * Override: Should only be used for development purposes (QA - Feature dev - e2e)
@@ -260,14 +259,9 @@ export namespace dynamicApi {
   function isEnabled(nameSpacedFlagName: string, defaultValue: boolean, context?: unknown): boolean;
 
   /**
-   * Getting string value of a string flag
+   * Getting string value of a Variant flag
    */
   function value(nameSpacedFlagName: string, defaultValue: string, context?: unknown): string;
-
-  /**
-   * Getting string value of a number flag
-   */
-  function getNumber(nameSpacedFlagName: string, defaultValue: number, context?: unknown): number;
 }
 
 export const flags: ReadonlyArray<Flag>;
